01_tasktopia/tasktopia-requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dash==2.9.3
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2024.5.2
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
django-plotly-dash==2.3.1
djangorestframework==3.15.1
dnspython==2.6.1
docopt==0.6.2
dpd_components==0.1.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
flashtext==2.7
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.52.4
fsspec==2024.6.0
funcs==0.10.1
google-auth==2.30.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gspread==6.1.2
gunicorn==22.0.0
h11==0.14.0
hdfs==2.7.3
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
llvmlite==0.42.0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
named==1.4.2
nest-asyncio==1.6.0
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
partd==1.4.2
patsy==0.5.6
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
psycopg2==2.9.9
py4j==0.10.9.7
pyahocorasick==2.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pynndescent==0.5.12
pyparsing==3.1.2
PySocks==1.7.1
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.1
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
selenium==4.19.0
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
solus==1.2.2
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
starlette==0.37.2
statsmodels==0.14.2
sudo==1.0.0
tenacity==8.3.0
threadpoolctl==3.5.0
toolz==0.12.1
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
typing-aliases==1.10.1
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
umap-learn==0.5.6
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.26.2
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
whitenoise==6.6.0
wraps==0.12.0
wsproto==1.2.0
zipp==3.19.2
